NoTransition scene config causes blank scenes
The NoTransition scene config produces an interpolator function with a
divide by zero, which can cause scenes to have opacity=NaN when
directionAdjustedProgress is set to 1.

This bad value is interpreted on Simulator as opacity=1 (bug does not
appear) but on my device it shows up as opacity=0 (scene shows briefly and
then disappears).

The 'constant' type seems more appropriate here.